Make Wallet public and usable in tests with ClientContext.
